[core] New leader key! Project leader key.
This new leader is a whole new scope for modes and key bindings in
Spacemacs.

The current scope are `global` and `major-mode`. The `major-mode`
scope has a dedicated leader `SPC m` or `,` as shortcut.

This commits adds a new scope: `project`. It hosts key bindings for
minor modes enabled for a project (in the projectile sense). This
was a missing piece in Spacemacs to correctly handle minor modes
like Ruby On Rails framework key bindings. These key bindings need
to be available in the whole project no matter the current major
mode.

The new scope has a new dedicated leader key: `SPC M` and `;` as
shortcut. `;` was chosen because it is the symmetrical of `,` in
VIM thus it makes sense to use it.

* new dotspacemacs variables:
    dotspacemacs-project-minor-mode-leader-key
    dotspacemacs-project-minor-mode-emacs-leader-key
* new functions in `core-keybindings.el`:
    spacemacs/declare-prefix-for-project-minor-mode
    spacemacs//init-project-minor-mode-leader-mode-map
    spacemacs/set-leader-keys-for-project-minor-mode
* new keyword in spacebind: :project-minor
* new command prefix: "M" for "Project commands"